small ajusting

+++ testing/arduino/self-driving-v3.ino
@@ -15,7 +15,6 @@ const uint8_t motorPin1_2 = 5;
 const uint8_t motorPin2_1 = 7;
 const uint8_t motorPin2_2 = 8;
 
-
 const uint8_t motorEnablePin = 3;
 
 const uint8_t rangeLimit_Stop = 40;
@@ -108,7 +107,7 @@ void loop()
 {
     long range = GetUltrasoundRange();
 
-    int motorSpeed = 1 * 255;
+    uint8_t motorSpeed = 1 * 255;
     uint32_t color = LEDStrip.Color(0, 255, 0);
     if(range < rangeLimit_Warning1) { motorSpeed = 0.6 * 255; color = LEDStrip.Color(127, 255, 0);}
     if(range < rangeLimit_Warning2) { motorSpeed = 0.5 * 255; color = LEDStrip.Color(255, 255, 0); }
